separated project into manageable files
