+++ src/config.h
@@ -1,0 +1,12 @@
+#ifndef CONFIG_H
+#define CONFIG_H
+
+#define DEFAULT_B_WIDTH 5
+#define DEFAULT_RADIUS 5
+#define DEFAULT_T_HEIGHT 25
+#define DEFAULT_U_COLOR 0x3f3f3f
+#define DEFAULT_F_COLOR 0x40E446
+#define DEFAULT_FONT "-*-terminus-*-*-*-*-12-*-*-*-*-*-*-*"
+#define NUM_TAGS 9
+
+#endif
