+++ src/types.h
@@ -1,0 +1,12 @@
+#ifndef STRUCTS
+#define STRUCTS
+/* std includes */
+#include <stdlib.h>
+#include <stdio.h>
+#include <stdbool.h>
+/*libraries */
+#include <X11/Xlib.h>
+
+
+
+#endif
