--- xcb/wm.h
@@ -1,10 +1,0 @@
-#ifndef WM_H
-#define WM_H
-
-#include <xcb/xcb.h>
-
-extern bool isrunning;
-
-static void run(void);
-
-#endif
